+++ tests/tine20/AllTests.php
@@ -37,8 +37,6 @@ class AllTests
             'ActiveSync',
             'Admin',
             'Courses',
-            'Voipmanager',
-            'Phone',
             'Timetracker',
             'Filemanager',
             'Projects',