+++ tine20/RELEASENOTES.md
@@ -2,8 +2,8 @@
 TINE RELEASENOTES
 =====================
                     
-  Release:     TBA (2024.11)
-  Last change: 2023-12-20
+  Release:     Pelle (2024.11)
+  Last change: 2024-07-24
 
 # GENERAL CHANGES (Administrative)
 
@@ -26,3 +26,7 @@ TINE RELEASENOTES
 # TASKS
 
 # HUMAN RESOURCES
+
+# PHONE & VOIPMANAGER
+
+- Both applications are discontinued and have been removed. Latest versions are in 2023.11 branch.